Added full folder interface for service values
The current `IFolder` interface didn't include all the relevant values from the service. This adds those values, and tests for them.

Existing tests were also beefed up to validate more strictly that the _thing_ that we did _actually_ hapened.

Note that this is not 100% of the fields returned by the service. Specifically `display_title` (seems to be duplicate of the actual `title` field), and `slug` are missing. It's unclear what slug actually *is*.