Remote Service State Cleanup in Test Setup
Because the API tests are attempting to run through some steps with *known* state on the service, it makes it a little difficult to *get* into that known state without an API wrapper.

Rather than contining down the path of very specific ordering of tests where we set let the tests mutate the state to a known place, this change implements a dumb API wrapper **just for the tests** to clean up the remote tests.

The tl;dr back story on this is that adds have a limit in the instapaper API, so you can't add more than N in 24hrs. By moving things to a clear state and minimizing the deletes, this helps make things cleaner.

With this API wrapper, we:
- Remove all folders (which moves the items in the folder to the archive)
- Unarchive all the bookmarks
- Remove liked state on any that are liked
- Reset read progress to 0.0

Additional changes
- Make internals of API library visible to tests to allow one one set of endpoint constants
- Removed the tests that aren't part of state setup from ordering