Code clean up
- Documentation for `IBookmark`
- Renamed missspelt variable
- Validate hash is returned when checking properties after an add

Note that not all properties are validated on the bookmark object when adding. Thi is due to a bug in the service when deleting a folder. Documentation says that deleting a folder will cause its contents to be moved to archive. They aren't, which means they can't be seen to be deleted so that 'add' is always fresh.

Rather than making more 'add' calls to get the items back into unread, we're just not going to validate the status of the progress or progress timestamp.