Add Support for List returning deleted items
When you List bookmarks, providing the Have information for _deleted_ bookmarks, the API will return a comma-separated list of the IDs that have been deleted from that folder -- in the meta object.

This requires changes to both the API and the shared helpers to support parsing that meta object. So:
- updated `List` to return a Tuple, with Bookmarks and Deleted IDs
- Updated `PerformRequestAsync` to return bookmarks and optional `JsonElement`
- Updated call callers to use these methods
- Updated all the tests