Folder Type + Add/Get of folders
This change adds the basic goo to be able to create a folder, and ask for it by explicit ID (service ID, or Local DB ID).

Also included are a number of extension methods to make working with some ADO.NET abstractions a little smoother.

The folders table was changed to allow `service_id` to be null due to locally-added folders that won't have serivce IDs.