Fix Instapaper API tests that broke after adding lazy properties.
In an earlier commit, we changed some properties to be lazy evaluated. However, we apparently forgot to run the unit tests, since they were 100% failing.

They were failing because the act of setting the user agent components cleared the Lazy<T> helper, causing it to always be null if you set the user agent.

This adds a helper method to be called from init/prop change, and now everything is peachy.