Remove `ulong` usage from the Folders service API
When working through some other implementations, it became clear that
typing the API surface as `ulong` to give tigther bounds on the API
service was ill-concieved.

JSON (well, JavaScript) only supports 53-bits for it's numeric type
(which is always floating point), and haveing 64bits in a ulong was
going to overflow more likely than not.

Switching to long is still not an exact match, but is a tighter mapping
to the real world.