Add folder sorting based on position
Folders from the service have an explicit position field that controls the relative ordering of the returned folders. We weren't respecting that.

As part of respecting that value, there are two problems:
1. What to do about Well Known Folders -- Unread, Archive
2. What to do about folders that are added but haven't been sync'd yet.

Solution is:
1. Well Known folders get negative positions
2. Unsync'd folders get position 0

For now, this means they're returned by the database w/ unsynced folders in the middle of the two sets.