Overwrite any local hash when updating read progress
The `hash` field is used by the Instapaper Service to determine if the read progress has changed. How this hash is generated is opaque to clients, so to ensure that the right thing happens during a sync, any chance made to progress needs to clear the hash.

This change does that when the *explicit* update of progress happens. Progress changes from the service will be handled differently through explicit updates to all fields