Add support for updating bookmarks
To support sync scenarios, we will need to be able to update all the fields on a bookmark. There was was no API for that, and this change adds it.

Note there is the additional change of removing nullability from the description field because it was making things messy, and the reality of the situation is that it won't be null.