Add the skeleton of loading local-only bookmark info
Local-only information for bookmarks is stored in a separate table to keep the
two concepts at least *logically* separated. This change adds the skeleton of
loading that information at the same time you get a bookmark (if that
information is present). Future changes will build out the
`DatabaseLocalBookmarkState` class.

Additionally:
- Renames `bookmark_with_download_state` to `bookmark_local_only_state` for
  clarity
- Updates unit tests to test that local-only information isn't returned if it's
  not present
- Exposes Local State on a `DatabaseBookmark`