Made `UpdateFolderAsync` descrption clearer
The Instapaper API doesn't allow clients to change the title of a folder.
However, the service itself does enable that functionality, so a client may see
changes from the service changing the title, and will need to update the
database.

Comment is clarifying that no *service* changes will come from this API.

Note, folder order *can* be changed, but isn't something we manage in the API