Added better handling of errors in the migration script
It's probably a little much to call it a 'migration' script, since it's
really a setup script at this point. *Anyway*, there have been a few
occurances where the script didn't report an error, but it hadn't run to
completion, causing failures.

To detect those, we leverage the fact we set a `user_version` on the
database once setup to detect if we successfully completed the script.