Add fail fast when API Keys are missing
Since we're not using explicitly included files in the project, if the
API key files are not created, you get an error about missing classes
and members. This isn't a very nice message and makes it feel like the
project is broken.

Now an error message will be displayed if you fail to create the
required files.