Implement Folder Ledgering
When instantiated, `Ledger` will now monitor the database for changes, and
create pending edits for folder operations. This includes resurrection of
folders seen by the service when a folder with the same title is recreated
before a sync is complete.

Tweaks: - `FolderAdded` need to be lowered to `string`, to ensure that the
handler has a chance to update the folder info before it's returned to the
folder creator - `UpdateFolder` was always creating a transaction; nested
transactions aren't supported, so we needed to be opportunistic in both creation
*and* cleanup