Extract events into internal-only interface.
The events as currently exposed are within the transaction of the code that
raises the event. This means that any exception thrown by handlers attached to
those events could cause the transaction to rollback. This is the beahviour we
want.

*However*, there are future listeners (e.g. the UI) that might throw exceptions
while handing not-related-to-data operations (E.g. button click). We don't want
those to stall forward progress.

Wrapping the handlers in try/catch was an option, but we do want to have some of
the internal handlers raise exceptions. Having the listeners themselves do the
hardwork also seemed wrong. Finally, we have the future consideration of not
wanting UI operations to delay the data mutations.

So we moved them out here to internal only interfaces.