Enable `STRICT` SQLite support
`STRICT` tables properly validate the data types of columns -- this is obviously
a desireable thing. This can be seen from the small fixes required here to
ensure the data types supplied to `CREATE TABLE` matched the types that the
SQLite wrapper actually writes to the database.

To do this, we had to decouple -- at least for now -- from the 'batteries
included' `Microsoft.Data.Sqlite` package, and use the 'core' variant. This
allowed to us to use a _newer_ version of the SQLitePCL wrapper package. (v6 of
the `Microsoft.Data.Sqlite` package used a < 3.37.0 version of SQLite that
doesn't support `STRICT`).

Maybe at somepoint we can return to the default batteries included version.