Remove interdependency between database access classes
There had grown a (sensible, justified) indeterdependency between the different
access classes -- to do X with class A, class A also needed access to class B &
C. This turned out to be mildly annoying as we look to do some heavy refactors
around the connection management. It was also clear the use cases were super
small, and maybe didn't justify it.

This change removes those interdependencies, and makes some small trade off in
less-specific exposures of the errors.