Factor away `IInstapaperDatabase`
The `IInstapaperDatabase` interface was intended to be the manager of all the
instances of database access layers. While it did satisfy this purpose, it
became clear as we move towards rethinking how connections exist/work that this
wasn't quite the right pattern. What the specific pattern we need is a little
unclear -- so lets remove this.

This allowed a significant clean up of the test classes to be more isolated and
only use the things they actually used.