Enhance “Within Transaction” Events with more data (Instance, Connection)
To loosen the coupling between components, the “WithinTransaction” events from
the data access layers have been tweaked to include additional information:
- The data access layer instance
- Connection

This allows listeners (e.g. the ledger) who are participating within the
transaction to operate on the same connection, and database explicitly rather
than hoping the one they've got is the right one.

As with the previous commits, this is to move the connection management into a
better pattern.