Make data access classes take a factory rather than connection
Continuing to make connections & interactions more ephemeral, to allow better
isolation of transactions et al.

With these changes, the public interface now creates a new connection for it's
actions. If the connection returned by the factory is *already opened*, then
the connection returned by the factory won't be closed or disposed, and it's
up to the callers to sort out the resulting behaviour.

This change also changes the ledger to no longer rely on the accessor class that
is supplied when the event raised (e.g. the first param), because that results in
connections being created, rather than reused. It likely needs to be removed
long term to prevent fat-fingering it.

Additionally, the connection is now set up as a shared cache in-memory
connection.