Dispose of ledgers correctly: Before doing a sync
In the real-world use case, the ledger is expected to exist on a separate DB
connection to the syncing connection, which would result in (correctly) sync
driven updates not creating pending changes in the database.

However, in the test scenario, to simplify it, we're re-using the connection (at
least for now). Thus, we need to disable ledgering as soon as we've constructed
the appropraite database state.