Test that unrelated article property changes are applied locally during sync
There are cases where we're syncing changes for a single aspect of an
article (E.g. move, like et al), but there are *other* changes service
side that we would like to pick up.

Rather than wait for them to be picked up in the intended sync phase, we
can leverage the service returning the updated state implicitly to
update the local store as we go.