Implement basic list-have integration
The primary mechanism for syncing is not explicit operations (moves,
likes et al), but the 'have' payload provided with folder listing
requests. This change implements the basic parts of that behaviour.

The most significant part of this change is a mock attempting to
implement the behaviour of the service. This is probably more logic than
one would like in in tests, but given the challenges of running against
the real service, it continues to feel like the right trade off.

Additional changes:
- Make some 'component' sync methods `internal`, rather than private.
  This is to simplify the tests, which were calling the 'uber' bookmark
  sync methods causing inconsitent behaviour 'cause of the larger
  scope of behaviour