Perform sync of liked-article folder
The liked folder is special case on the service -- articles are in a
folder, but they are _implicitly_ in the "liked" folder if they've been
liked. When syncing, if we haven't enumerated all the folders, there
might be like-state changes that we haven't seen. This means we need to
do a 'mop up' sync of the virtual liked folder.

This change adds the basic support for that, with a slight refactor of
the single folder syncing to pull some reusable parts out.

Includes basic unit tests.