Implement basic per-folder sync limits + orphan cleanup
Folder sync limits impact the syncing likes. Ignoring likes, it's pretty
simple: Gather deletes, do deletes.

With likes, there are articles in the virtual like folder that *aren't*
inside the limit of any other folder, and thus wouldn't be available.
But the expectation is that they should be. This means special
consideration for likes is needed -- esp. wrt cleaning up orphans. Hence
why they're both included.

Specifically, orphan clean up needs to respect the folder limit *for the
like folder*. If there are orphaned likes, *but* they're under the limit
for likes, they shouldn't be purged. But if they are, they should be.

This change adds the basic support -- basic because at a future point
we'll need to add *per folder* limits.