Add Transaction Rollback test for Sync
Because we would like to use a single-large transaction to allow rolling
back the local database if the sync fails, we need to test that it
works.

The test here completes a sync -- and ensures things match -- and then
rolls the transaction back, and verifies there was, in fact, some data
still present (e.g. isn't empty)