Add Service Integration Sync Tests
Add scoped, small, tests that perform a real sync with the service. This
is to make sure we're not missing any quirky service behaviours.

As part of implementing this, we exposed a deficiency in the testing for
handling a service bug, against empty states. The documentation says
articles in folders will be moved to the archive folder when the
containing folder is deleted. In actuality, they aren't placed anywhere
(the bug). This introduces a scenario where liked articles now *only*
appear in the liked 'virtual' folder.

To address this, the tests are adjusted to mimic that scenario in the
mock service implementation, as well as the state configuration of the
test database.

This resulted in a bug in our handling of syncing implicit like-related
changes on the service -- we now add the article if it's not otherwise
found in our database.

Other things of note:
- Take a dependency on the `InstapaperAPITests` project for the wrapper
  and state configuration of the service. No point in duplicating, and
  it felt cleaner to add them here -- and the advantage of making it
  easier to open the API tests without a dependency on the rest of
  storyvoid
- Add diagnostics to the Sync tests ('cause the service wrapper logs a
  bunch of stuff)