Move clearing house implementations into test projects
The database & sync engines don't need to have a concrete implementation
of the event sink & source. This is because the consumer is likely able
to provide better marshalling of the events to the right places, rather
than having them codified in the database implementation, and the
consumer _effectively_ reimplementing the interface anyway to marshall.

(Also renames `IContentSyncEvent*` to `IDatabaseSyncEvent*` for clarity.
It wasn't really about content sync, but the database sync)