Add File type detection + first image selection
There are some image references that don’t have an extension in the file name. We can still render them, but we’d like a proper file extension for them.

We can use ImageSharp to detect the format, and set the file extension based on that.

Additionally, since we now know the size of the image (thanks ImageSharp!), we can apply filtering so we don’t pick small (<150px) images for our first image — which will be used as a thumbail.