Create progress events for article downloading
We need to support progress display in the UI, and we use events to
expose that information. This change adds the basic clearing house and
test for it.

Along with that it became clear there were some missing error cases that
this change made behavioural changes to (missing articles), so those
were also added.

The other motivation for this change is to allow future tests to delay
or block progress being made to facilitate cancellation testing of long
operations.