Implement intial change processor for articles
To apply changes from the database to the visible list, we need a component to listen for those changes, and make the matching change to the list used for display.

This would be trivial, except that list is sorted, and the items need to be placed in the correct location, not just at the end.

With the component added here, adds + updates to articles are processed & inserted into a list.

It is not yet included in the `ArticleList` view model.