Refactored `ArticleCommand` to operate on `DatabaseArticles`
As the more complex commands (Archive, Like/Dislike) are implemented, it
became clear that we needed more than IDs. We'd need to query the
database to determin enabled state. This seemed expensive for something
that we *already* had the information in the `DatabaseArticle` instance
itself.

So, switched it up to use those instead. Hopefully this pans out.