Concurrent initialization of the database
This creates an (internal) method for obtaining the database on `AppUtilities`, which is supplied with a task to do that by the app.

This is to ensure that the UI can be brought up without waiting for the DB to completed first *or* on demand when we need it.

Some of the code here is still placeholder while we continue the transition to a 'production' navigation system