Fix Code Analysis DLL load errors in App solution
When opening the `-WinApp` solution, there were warnings reported from code analysis being unable to load an assembly.

Turns out this is because the source generator (Reswinator) used by the UI project references a set of assemblies that aren't compatible with the default ones. By explicitly declaring the dependencies, things are fixed.

Still unknown: Why it's InstapaperAPITests that experiences this issue.