Reference App project + include test to validate
This change got weird. With the packaged app, the XAML generated code in the **Test** project couldn't locate the XBF that backed the markup.

But, we didn't really need that once I thought about it. Removing that allowed the test harness stuff to locate tests & execute.

Also exposed the window created to the tests so that we can use it with the utilities class (it needs a `Frame`