Add tests + test support for the `AppUtilities` class
This change adds tests for the async nature of database opening.

Of note, this change includes the infrastrrcture to switch to the UI thread in a test.

Why not just use `UITestMethod`? Well, that doesn't allow `async` test methods, so you have to write `TestMethod`, enqueue to the UI thread, and do the work there.

This ultimatley seemed more sensible than doing a giant lambda, and felt natural