Fix deadlock caused by disposal of `AppUtilities`
I was too clever by half on the cleanup of the connection etc in `AppUtilities`. I tried to avoid the asynchronicity of disposable with a `Wait`. This was a bad idea, since we did a DB Clean up on 'sign out' in `MainWindow`, so we deadlocked on the UI thread disposing the task, which was bad.

While this is *still* broken, we don't need to actually clean up -- it's not quite right now anyway. So, lets just remove the cleanup.