Make `DownloadArticleAsync` concurrent download aware
To allow a more 'fire and forget' behaviour when trying to download
articles, we need `DownloadArticleAsync` to only initiate the download
once if a second request is made before the first completes.

To do this, we're going to track the downloads by `Task`, and if we have
a task for a given article, return that instead.

This includes tests to verify the behaviour & eventing.