Make database sync single-op-at-a-time
Previously, when a sync was started, off it went. It was possible — for a given instance of `InstapaperSync` to initate over & over while previous syncs were in progress. This was probably bad, but didn’t really impact tests.

As we come to adding syncing to the main app, it became clear that we needed to minimize the window where this might happen with the UI allowing things to happen ‘quickly’.

Rather than shove this into the UI layer, it felt more sensible to push it down here.

As part of building out the tests we learned that .net runtimes task execution will use the *same* completed task instance for void tasks if they’re already completed (neat, sensible optimization). This however made validation of the logic here troublesome.

So, as well as implementing the behaviour, added the ability to have a sync actually complete asynchronously by adding a real delay to the folder mock (disabled by default).