Create `SyncHelper` class to manage sync + database connection
Syncing is complicated due to the need for an transactionally-isolated
database instance. We also want to aggregate both the database sync,
with the article downloads when performing a sync. This means we need
something to tie them together.

`SyncHelper` aggregates the operation into a cancellable class that
allows consumers to listen for state changes.

Tests included. Of note, interfaces were created for `ArticleDownloader`
and `InstaperSync` so that `SyncHelper` could be tested in isolation
without needing to spin up a full set of databases for syncing and
downloading.