Consume `SyncCommand` from Article List Page
This adds a new parameter to the `ArticleList` that is passing through the sync command to the page. That command is then added to the toolbar flyout.

Note that because of the addition to the `ArticleList` constructor, the mock sync items are moved from `SynCommand.Tests.cs` to their own file.