Remove `AngleSharp.Css` -- behaviour moved to the main `AngleSharp` library
When originally written, there was no support for accessing the text nodes of the DOM tree in an appropriate manner _in the core AngleSharp library_. However it was in `AngleSharp.Css`.

In the intervening 3 years, the core library gained the appropriate DOM-equiv. api, allowing access to the text of the subtree for generating the subscription.

This updates to that newer API. Note the additional `Trim()` is because the whitespace isn't trimmed (e.g. \n)