Fix issue when signing in a second time in the same session + fixing/adding test for scenario
We weren't cleaning up the db task correctly, so when we accessed the data layer a second time, we didn't get the new one -- just the old, disposed of one.

Note, to add a test for this scenario required factoring out some of the UI Navigation from the `AppUtilities` class. Not a fan of the indirection, but it ended up being clean enough with a 'delegate' pattern